cloudFunctions/firestoreLogging/requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
functions-framework==3.*
google-cloud-firestore==2.13.1
